Split State/StateRead in v1/v2
