+++ fractal_server/app/schemas/v2/__init__.py
@@ -1,3 +1,4 @@
+from .collection_state import CollectionStateReadV2  # noqa F401
 from .dataset import DatasetCreateV2  # noqa F401
 from .dataset import DatasetExportV2  # noqa F401
 from .dataset import DatasetImportV2  # noqa F401
