+++ tests/v1/02_schemas/test_unit_schemas.py
@@ -5,8 +5,6 @@ import pytest
 from devtools import debug
 from pydantic.error_wrappers import ValidationError
 
-from fractal_server.app.schemas.state import _StateBase
-from fractal_server.app.schemas.state import StateRead
 from fractal_server.app.schemas.user import UserCreate
 from fractal_server.app.schemas.user import UserUpdate
 from fractal_server.app.schemas.user import UserUpdateStrict
@@ -21,6 +19,7 @@ from fractal_server.app.schemas.v1 import ProjectCreateV1
 from fractal_server.app.schemas.v1 import ProjectReadV1
 from fractal_server.app.schemas.v1 import ResourceCreateV1
 from fractal_server.app.schemas.v1 import ResourceReadV1
+from fractal_server.app.schemas.v1 import StateRead
 from fractal_server.app.schemas.v1 import TaskCollectPipV1
 from fractal_server.app.schemas.v1 import TaskCreateV1
 from fractal_server.app.schemas.v1 import TaskImportV1
@@ -322,10 +321,6 @@ def test_project_create():
     # Fail due to empty string
     with pytest.raises(ValidationError):
         ProjectCreateV1(name="  ")
-
-
-def test_state():
-    _StateBase(data={"some": "thing"}, timestamp=get_timestamp())
 
 
 def test_state_read():
